use nscd user if we find it, nobody if not...
this should make the change backcompat -- if there is an nscd user we'll
use it, if not we won't.

note that due to glibc caching (and since we're managing nscd) on first
provisioning this may miss the creation of the nscd user and get into
a chicken/egg problem.  the best solution there is not to bother
creating an nscd user and to just use 'nobody'.